+++ src/game/battlefield/mod.rs
@@ -4,6 +4,7 @@
 
 pub(in crate::game) mod components;
 pub(in crate::game) mod constants;
+pub(in crate::game) mod flora;
 mod plugin;
 pub(in crate::game) mod systems;